backend/requirements.txt
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
google-cloud-vision>=3.0.0
